requirements.txt
paho-mqtt
rumps 